refactor(examples): extract token panel into its own component file
Move the inline SelectedTokenPanel (TokenWorkspace + NoTokenWorkspace in
react-wagmi) out of each app's page.tsx into a dedicated component file, so
page.tsx is just the connect/switch/header/token-selector shell. Pure move —
no logic, prop, or call-site changes; each new file carries only the imports
its component uses and page.tsx drops the now-unused card/query-key imports.

All example apps typecheck clean.
