+++ examples/react-ethers/src/components/SelectedTokenPanel.tsx
@@ -1,0 +1,113 @@
+"use client";
+
+import { useQueryClient } from "@tanstack/react-query";
+import type { Address, TokenWrapperPairWithMetadata } from "@zama-fhe/sdk";
+import { zamaQueryKeys } from "@zama-fhe/sdk/query"; // query key builders for SDK-managed caches — /query subpath export
+import { erc20BalanceKey } from "@/lib/queryKeys";
+import { BalancesCard } from "@/components/BalancesCard";
+import { ShieldCard } from "@/components/ShieldCard";
+import { TransferCard } from "@/components/TransferCard";
+import { UnshieldCard } from "@/components/UnshieldCard";
+import { PendingUnshieldCard } from "@/components/PendingUnshieldCard";
+import { DelegateDecryptionCard } from "@/components/DelegateDecryptionCard";
+import { RevokeDelegationCard } from "@/components/RevokeDelegationCard";
+import { DecryptAsCard } from "@/components/DecryptAsCard";
+
+interface SelectedTokenPanelProps {
+  address: Address;
+  token: TokenWrapperPairWithMetadata;
+  validPairs: TokenWrapperPairWithMetadata[];
+  isSepolia: boolean;
+  ethBalanceKey: readonly unknown[];
+}
+
+export function SelectedTokenPanel({
+  address,
+  token,
+  validPairs,
+  isSepolia,
+  ethBalanceKey,
+}: SelectedTokenPanelProps) {
+  const queryClient = useQueryClient();
+
+  const refreshBalances = () => {
+    queryClient.invalidateQueries({ queryKey: erc20BalanceKey(token.tokenAddress, address) });
+    queryClient.invalidateQueries({ queryKey: ethBalanceKey });
+    // Invalidate the encrypted handle so useConfidentialBalance re-polls after
+    // any operation that changes the confidential balance (shield, unshield, transfer).
+    queryClient.invalidateQueries({
+      queryKey: zamaQueryKeys.confidentialBalance.token(token.confidentialTokenAddress),
+    });
+  };
+
+  const actionsDisabled = !isSepolia;
+
+  return (
+    <>
+      <BalancesCard
+        token={token}
+        account={address}
+        validPairs={validPairs}
+        disabled={actionsDisabled}
+        onSuccess={refreshBalances}
+      />
+
+      {/* Pending unshield resume — checked for every registered token, not just the selected one.
+          key includes address so the component remounts (re-checks IndexedDB) on wallet change. */}
+      {validPairs.map((pair) => (
+        <PendingUnshieldCard
+          key={`${pair.confidentialTokenAddress}-${address}`}
+          token={pair}
+          onSuccess={refreshBalances}
+        />
+      ))}
+
+      <h2 className="section-label">Operations</h2>
+
+      {/* key includes address and selected token so cards remount (inputs + state reset) on wallet or token change */}
+      <ShieldCard
+        key={`shield-${address}-${token.confidentialTokenAddress}`}
+        token={token}
+        disabled={actionsDisabled}
+        onSuccess={refreshBalances}
+      />
+
+      <TransferCard
+        key={`transfer-${address}-${token.confidentialTokenAddress}`}
+        token={token}
+        disabled={actionsDisabled}
+        onSuccess={refreshBalances}
+      />
+
+      <UnshieldCard
+        key={`unshield-${address}-${token.confidentialTokenAddress}`}
+        token={token}
+        disabled={actionsDisabled}
+        onSuccess={refreshBalances}
+      />
+
+      <h2 className="section-label">Delegation — as owner</h2>
+
+      <DelegateDecryptionCard
+        key={`grant-delegation-${address}-${token.confidentialTokenAddress}`}
+        token={token}
+        disabled={actionsDisabled}
+      />
+
+      <RevokeDelegationCard
+        key={`revoke-delegation-${address}-${token.confidentialTokenAddress}`}
+        token={token}
+        disabled={actionsDisabled}
+      />
+
+      <h2 className="section-label">Delegation — as delegate</h2>
+
+      <DecryptAsCard
+        key={`decrypt-as-${address}-${token.confidentialTokenAddress}`}
+        token={token}
+        account={address}
+        disabled={actionsDisabled}
+      />
+    </>
+  );
+}
