+++ examples/react-viem/src/app/page.tsx
@@ -6,16 +6,7 @@ import { formatEther, createPublicClient, http } from "viem";
 import { sepolia } from "viem/chains";
 import { useListPairs } from "@zama-fhe/react-sdk";
 import type { Address, TokenWrapperPairWithMetadata } from "@zama-fhe/sdk";
-import { zamaQueryKeys } from "@zama-fhe/sdk/query"; // query key builders for SDK-managed caches — /query subpath export
-import { erc20BalanceKey } from "@/lib/queryKeys";
-import { BalancesCard } from "@/components/BalancesCard";
-import { ShieldCard } from "@/components/ShieldCard";
-import { TransferCard } from "@/components/TransferCard";
-import { UnshieldCard } from "@/components/UnshieldCard";
-import { PendingUnshieldCard } from "@/components/PendingUnshieldCard";
-import { DelegateDecryptionCard } from "@/components/DelegateDecryptionCard";
-import { RevokeDelegationCard } from "@/components/RevokeDelegationCard";
-import { DecryptAsCard } from "@/components/DecryptAsCard";
+import { SelectedTokenPanel } from "@/components/SelectedTokenPanel";
 import {
   SEPOLIA_CHAIN_ID,
   SEPOLIA_CHAIN_ID_HEX,
@@ -56,103 +47,6 @@ async function switchToSepolia(ethereum: NonNullable<ReturnType<typeof getEthere
     // wallet_switchEthereumChain errors other than 4902 (including 4001 rejection) are
     // intentionally ignored — chainId is re-read in the finally block of the caller.
   }
-}
-
-interface SelectedTokenPanelProps {
-  address: Address;
-  token: TokenWrapperPairWithMetadata;
-  validPairs: TokenWrapperPairWithMetadata[];
-  isSepolia: boolean;
-  ethBalanceKey: readonly unknown[];
-}
-
-function SelectedTokenPanel({
-  address,
-  token,
-  validPairs,
-  isSepolia,
-  ethBalanceKey,
-}: SelectedTokenPanelProps) {
-  const queryClient = useQueryClient();
-
-  const refreshBalances = () => {
-    queryClient.invalidateQueries({ queryKey: erc20BalanceKey(token.tokenAddress, address) });
-    queryClient.invalidateQueries({ queryKey: ethBalanceKey });
-    queryClient.invalidateQueries({
-      queryKey: zamaQueryKeys.confidentialBalance.token(token.confidentialTokenAddress),
-    });
-  };
-
-  const actionsDisabled = !isSepolia;
-
-  return (
-    <>
-      <BalancesCard
-        token={token}
-        account={address}
-        validPairs={validPairs}
-        disabled={actionsDisabled}
-        onSuccess={refreshBalances}
-      />
-
-      {/* Pending unshield resume — checked for every registered token, not just the selected one.
-          key includes address so the component remounts (re-checks IndexedDB) on wallet change. */}
-      {validPairs.map((pair) => (
-        <PendingUnshieldCard
-          key={`${pair.confidentialTokenAddress}-${address}`}
-          token={pair}
-          onSuccess={refreshBalances}
-        />
-      ))}
-
-      <h2 className="section-label">Operations</h2>
-
-      {/* key includes address and selected token so cards remount (inputs + state reset) on wallet or token change */}
-      <ShieldCard
-        key={`shield-${address}-${token.confidentialTokenAddress}`}
-        token={token}
-        disabled={actionsDisabled}
-        onSuccess={refreshBalances}
-      />
-
-      <TransferCard
-        key={`transfer-${address}-${token.confidentialTokenAddress}`}
-        token={token}
-        disabled={actionsDisabled}
-        onSuccess={refreshBalances}
-      />
-
-      <UnshieldCard
-        key={`unshield-${address}-${token.confidentialTokenAddress}`}
-        token={token}
-        disabled={actionsDisabled}
-        onSuccess={refreshBalances}
-      />
-
-      <h2 className="section-label">Delegation — as owner</h2>
-
-      <DelegateDecryptionCard
-        key={`grant-delegation-${address}-${token.confidentialTokenAddress}`}
-        token={token}
-        disabled={actionsDisabled}
-      />
-
-      <RevokeDelegationCard
-        key={`revoke-delegation-${address}-${token.confidentialTokenAddress}`}
-        token={token}
-        disabled={actionsDisabled}
-      />
-
-      <h2 className="section-label">Delegation — as delegate</h2>
-
-      <DecryptAsCard
-        key={`decrypt-as-${address}-${token.confidentialTokenAddress}`}
-        token={token}
-        account={address}
-        disabled={actionsDisabled}
-      />
-    </>
-  );
 }
 
 export default function Home() {
